# Conflicts:
#	backend/sendmessage/__tests__/index.test.js

+++ mobile/lib/controllers/ChatPageController.dart
@@ -2,12 +2,14 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:mobile/dialogs/ForwardDialog.dart';
 import 'package:mobile/domain/Chat.dart';
 import 'package:mobile/domain/Contact.dart';
 import 'package:mobile/domain/Message.dart';
 import 'package:mobile/models/ChatPageModel.dart';
+import 'package:mobile/pages/ChatPage.dart';
 import 'package:mobile/services/ChatCacheService.dart';
 import 'package:mobile/services/ChatService.dart';
 import 'package:mobile/services/CommunicationService.dart';
@@ -31,8 +33,9 @@ class ChatPageController {
   final String chatId;
   late final Future<Chat> chat;
   late final String contactPhoneNumber;
+  bool privateChatAccepted = false;
 
-  ChatPageController({required this.chatId}) {
+  ChatPageController({required this.chatId, Chat? privateChat}) {
     communicationService.onMessage(_onMessage);
 
     communicationService.onDelete(_onDelete);
@@ -43,10 +46,35 @@ class ChatPageController {
 
     communicationService.onMessageLiked(_onMessageLiked);
 
+    communicationService.onMessageEdited(_onMessageEdited);
+
     communicationService.shouldBlockNotifications =
         (senderPhoneNumber) => senderPhoneNumber == contactPhoneNumber;
 
-    chat = chatService.fetchById(chatId);
+    if (privateChat != null) {
+      chat = Future.value(privateChat);
+      communicationService.onAcceptPrivateChat = () async {
+        privateChatAccepted = true;
+
+        final chat = await this.chat;
+
+        final name = chat.contact?.displayName ?? chat.contactPhoneNumber;
+
+        model.addMessage(
+          Message(
+            id: Uuid().v4(),
+            chatId: chatId,
+            isIncoming: true,
+            otherPartyPhoneNumber: contactPhoneNumber,
+            contents: "$name has accepted the private chat",
+            timestamp: DateTime.now(),
+          ),
+        );
+      };
+    } else {
+      chat = chatService.fetchById(chatId);
+    }
+
     chat.then((chat) {
       contactPhoneNumber = chat.contactPhoneNumber;
       model.contactTitle = chat.contact?.displayName ?? chat.contactPhoneNumber;
@@ -95,7 +123,13 @@ class ChatPageController {
   void _onMessage(Message message) {
     if (message.chatId == chatId) {
       model.addMessage(message);
-      messageService.setMessageReadReceipt(message.id, ReadReceipt.seen);
+      communicationService.sendAckSeen([message.id], contactPhoneNumber);
+      chat.then((chat) {
+        if (chat.chatType == ChatType.general)
+          messageService.setMessageReadReceipt(message.id, ReadReceipt.seen);
+      });
+    } else if (message.otherPartyPhoneNumber == contactPhoneNumber) {
+      model.addMessage(message);
       communicationService.sendAckSeen([message.id], contactPhoneNumber);
     }
   }
@@ -128,14 +162,21 @@ class ChatPageController {
     model.setLikedById(messageID, liked);
   }
 
+  void _onMessageEdited(String messageID, String message) {
+    model.setEditedById(messageID, message);
+  }
+
   void dispose() {
     communicationService.disposeOnMessage(_onMessage);
     communicationService.disposeOnDelete(_onDelete);
     communicationService.disposeOnAck(_onAck);
     communicationService.disposeOnAckSeen(_onAckSeen);
+    communicationService.disposeOnMessageLiked(_onMessageLiked);
+    communicationService.disposeOnMessageEdited(_onMessageEdited);
     communicationService.shouldBlockNotifications = (phoneNumber) => false;
     contactService.disposeOnChanged(_onContactChanged);
     memoryStoreService.disposeOnContactOnline(_onOnline);
+    //communicationService.onStopPrivateChat = null;
   }
 
   void sendMessage(String contents) async {
@@ -150,9 +191,36 @@ class ChatPageController {
 
     final chatType = (await chat).chatType;
 
-    communicationService.sendMessage(message, chatType, contactPhoneNumber);
-    messageService.insert(message);
+    communicationService.sendMessage(
+        message, chatType, contactPhoneNumber, null);
     model.addMessage(message);
+    chat.then((chat) {
+      if (chat.chatType == ChatType.general) messageService.insert(message);
+    });
+  }
+
+  void replyToMessage(
+    String contents,
+    String? repliedMessageId,
+  ) async {
+    final message = Message(
+      id: Uuid().v4(),
+      chatId: chatId,
+      isIncoming: false,
+      otherPartyPhoneNumber: contactPhoneNumber,
+      contents: contents.trim(),
+      timestamp: DateTime.now(),
+      repliedMessageId: repliedMessageId,
+    );
+
+    final chatType = (await chat).chatType;
+
+    communicationService.sendMessage(
+        message, chatType, contactPhoneNumber, repliedMessageId);
+    model.addMessage(message);
+    chat.then((chat) {
+      if (chat.chatType == ChatType.general) messageService.insert(message);
+    });
   }
 
   void sendImage(Uint8List imageBytes) async {
@@ -166,34 +234,46 @@ class ChatPageController {
       isMedia: true,
     );
 
+    final chatType = (await chat).chatType;
+
     model.addMessage(message);
-    messageService.insert(message);
     communicationService.sendImage(
       message,
-      ChatType.general,
+      chatType,
       contactPhoneNumber,
     );
+    chat.then((chat) {
+      if (chat.chatType == ChatType.general) messageService.insert(message);
+    });
   }
 
   void deleteMessagesLocally(List<String> ids) {
-    ids.forEach((id) {
-      messageService.deleteById(id);
-      model.removeMessageById(id);
+    chat.then((chat) {
+      ids.forEach((id) {
+        model.removeMessageById(id);
+        if (chat.chatType == ChatType.general) messageService.deleteById(id);
+      });
     });
   }
 
   void deleteMessagesRemotely(List<String> ids) {
-    ids.forEach((id) {
-      communicationService.sendDelete(id, contactPhoneNumber);
-      messageService.setMessageDeleted(id);
-      model.setDeletedById(id);
+    chat.then((chat) {
+      ids.forEach((id) {
+        communicationService.sendDelete(id, contactPhoneNumber);
+        model.setDeletedById(id);
+        if (chat.chatType == ChatType.general)
+          messageService.setMessageDeleted(id);
+      });
     });
   }
 
   void likeMessage(String messageId, bool liked) {
     communicationService.sendLiked(messageId, liked, contactPhoneNumber);
-    messageService.setMessageLiked(messageId, liked);
     model.setLikedById(messageId, liked);
+    chat.then((chat) {
+      if (chat.chatType == ChatType.general)
+        messageService.setMessageLiked(messageId, liked);
+    });
   }
 
   void addSenderAsContact(String displayName) {
@@ -204,17 +284,22 @@ class ChatPageController {
       profileImage: "",
     );
 
-    contactService.insert(contact);
-
     communicationService.sendRequestProfileImage(contactPhoneNumber);
     communicationService.sendRequestStatus(contactPhoneNumber);
 
     model.contactSaved = true;
     model.contactTitle = displayName;
+
+    chat.then((chat) {
+      if (chat.chatType == ChatType.general) contactService.insert(contact);
+    });
   }
 
   void storeTypedMessage(String message) {
-    chatCacheService.put(chatId, message);
+    chat.then((chat) {
+      if (chat.chatType == ChatType.general)
+        chatCacheService.put(chatId, message);
+    });
   }
 
   String getTypedMessage() {
@@ -256,10 +341,51 @@ class ChatPageController {
           );
 
           communicationService.sendMessage(
-              newMessage, ChatType.general, element.first);
+              newMessage, ChatType.general, element.first, null);
           messageService.insert(newMessage);
         });
       });
     });
+  }
+
+  void editMessage(String messageId, String newMessage) {
+    communicationService.sendEditedMessage(
+        messageId, newMessage, contactPhoneNumber);
+    model.setEditedById(messageId, newMessage);
+    chat.then((chat) {
+      if (chat.chatType == ChatType.general)
+        messageService.updateMessageContents(messageId, newMessage);
+    });
+  }
+
+  void startPrivateChat(BuildContext context) async {
+    communicationService.sendStartPrivateChat(contactPhoneNumber);
+    Contact? contact;
+    try {
+      contact = await contactService.fetchByPhoneNumber(contactPhoneNumber);
+    } on ContactWithPhoneNumberDoesNotExistException {
+      contact = null;
+    }
+
+    final Chat chat = Chat(
+      id: Uuid().v4(),
+      contactPhoneNumber: contactPhoneNumber,
+      chatType: ChatType.private,
+      contact: contact,
+    );
+    Navigator.push(
+      context,
+      MaterialPageRoute(
+        builder: (_) => ChatPage.privateChat(
+          privateChat: chat,
+          chatId: chat.id,
+        ),
+      ),
+    );
+  }
+
+  void stopPrivateChat() {
+    if (privateChatAccepted)
+      communicationService.sendStopPrivateChat(contactPhoneNumber);
   }
 }
